+++ src/index.html
@@ -23,8 +23,9 @@
     <include src="./partrials/pagination.html"></include>
     <include src="./partrials/footer.html"></include>
     <include src="./partrials/modal.html"></include>
-
-    <!-- <div id="arrowTop" hidden></div> -->
     <include src="./partrials/visit-modal.html"></include>
+    <!-- <div id="arrowTop" hidden></div> -->
+    
+    <script type="module" src="./js/visitEvents.js"></script>
   </body>
 </html>
