+++ src/js/pageInStorage.js
@@ -1,28 +1,11 @@
 import storage from './storage'
 
-const STORAGE_PAGE_KEY = "last visit page";
-const storagePage = storage.load(STORAGE_PAGE_KEY);
-const refs = {
-    comeBackButton: document.querySelector('button.js-visit-back-btn'),
-    visitModal: document.querySelector("[data-modal].dialog"),
-    visitClose: document.querySelector("button.js-dialog-close"),
-    visitCloseBtn: document.querySelector("button.js-visit-close-btn"),
-};
+export const STORAGE_PAGE_KEY = "last visit page";
+export const storagePage = storage.load(STORAGE_PAGE_KEY);
 
-// currentPage = 8;
-
+currentPage = 2;
+ 
 changeStoragePage();
-
-(() => {
-        if (currentPage === 1 || storagePage === undefined) {
-            return
-     }toggleModal();
-})();    
-
-
-refs.comeBackButton.addEventListener("click", loadStoragePage); 
-refs.visitClose.addEventListener("click", toggleModal);
-refs.visitCloseBtn.addEventListener("click", toggleModal);
 
 function changeStoragePage() {
     if (currentPage > 1 && currentPage !== storagePage.value) {        
@@ -30,10 +13,5 @@ function changeStoragePage() {
         storage.save(STORAGE_PAGE_KEY, { value: currentPage });
     }    
 };
-function loadStoragePage() {     
-    if (currentPage !== 1){ currentPage = storagePage.value; } 
-    toggleModal();
-};
-function toggleModal() {   
-        refs.visitModal.classList.toggle("is-hidden");  
-};
+
+
