+++ src/partrials/visit-modal.html
@@ -7,8 +7,6 @@
     </div>
 
     <div class="dialog__button-wrap">
-        <button type="button" class="modal-movie__watched js-visit-back-btn">COME BACK</button>
-        <button type="button" class="modal-movie__queue js-visit-close-btn">CLOSE</button>
+        <button type="button" class="dialog__button  js-visit-back-btn">COME BACK</button>
+        <button type="button" class="dialog__button  js-visit-close-btn">CLOSE</button>
     </div>
-   </div> 
-</div>
