+++ src/sass/components/_button.scss
@@ -40,3 +40,25 @@
     width: 30px;
     height: 30px;
 }
+
+.dialog__button{
+color: $dark-text-color;
+    background-color: $main-bg-color;
+
+    font-weight: 500;
+    font-size: 12px;
+    line-height: calc(16 / 12);
+
+    width: 110px;
+    height: 44px;
+    border-radius: 5px;
+
+    border: 2px solid black;
+    &:hover,
+        &:focus {
+            border: none;
+            color: $header-text-color;
+            background: $accent-orange;
+            box-shadow: 0px 8px 43px rgba(255, 107, 1, 0.6);
+        }
+}
