+++ src/sass/components/_visit-modal.scss
@@ -1,15 +1,11 @@
 .dialog{
-    position: fixed;
-    z-index: 1;
+    position: fixed;    
     top: 0;
-    padding: 20px;
-    width: auto;
-    height: auto;
-    background-color: rgba(0, 0, 0, 0.2);
 }
 .dialog__container{
-    position:fixed;  
-    bottom: 50px;  
+    position:fixed;
+    left: 20;
+    bottom: 20px;  
     z-index: 100;
     padding: 20px;
     text-align: center;
